+++ modelopt/onnx/quantization/int8.py
@@ -27,7 +27,6 @@ from onnx_graphsurgeon.ir.node import Node
 from onnxruntime.quantization import CalibrationMethod
 from onnxruntime.quantization.calibrate import CalibrationDataReader
 
-from modelopt.onnx.autocast.convert import convert_to_f16
 from modelopt.onnx.logging_config import configure_logging, logger
 from modelopt.onnx.quantization.calib_utils import import_scales_from_calib_cache
 from modelopt.onnx.quantization.graph_utils import (
@@ -51,6 +50,7 @@ from modelopt.onnx.quantization.partitioning import (
     find_quantizable_nodes,
     get_skipped_output_layers,
 )
+from modelopt.onnx.quantization.precision_utils import _convert_to_runtime_precision
 from modelopt.onnx.quantization.qdq_utils import has_qdq_nodes, replace_scale_values
 
 
@@ -300,19 +300,16 @@ def quantize(
         if calibration_cache_path:
             replace_scale_values(onnx_model.graph, act_scales_dict)
 
-    if high_precision_dtype in ["fp16", "bf16"]:
-        # We need to convert float to float16 so as to speed up layers like LayerNorm or GroupNorm.
-        logger.info(f"Converting float32 tensors to {high_precision_dtype}")
-        # Note: from convert_to_f16's perspective, high_precision_dtype is the precision to reduce to from FP32
-        onnx_model = convert_to_f16(
-            onnx_model,
-            keep_io_types=not direct_io_types,
-            op_block_list=op_types_to_exclude_fp16 or [],
-            tensor_block_dict=custom_ops_to_cast_fp32 or {},
-            low_precision_type=high_precision_dtype,
-            trt_plugins=trt_extra_plugin_lib_paths,
-            opset=opset,
-        )
+    onnx_model = _convert_to_runtime_precision(
+        onnx_model,
+        quantize_mode="int8",
+        high_precision_dtype=high_precision_dtype,
+        direct_io_types=direct_io_types,
+        op_types_to_exclude_fp16=op_types_to_exclude_fp16,
+        custom_ops_to_cast_fp32=custom_ops_to_cast_fp32,
+        trt_extra_plugin_lib_paths=trt_extra_plugin_lib_paths,
+        opset=opset,
+    )
 
     if nodes_to_quantize:
         logger.info(f"Quantization completed successfully in {time.time() - t_start} seconds")